refactor; add metric formatting & icon to metric_view
