+++ webapp/src/components/backstage/playbook_edit/metrics/metric_edit.tsx
@@ -4,13 +4,13 @@
 import React, {useState} from 'react';
 import styled, {css} from 'styled-components';
 import {useIntl} from 'react-intl';
-import {Duration} from 'luxon';
 
 import {Metric, MetricType} from 'src/types/playbook';
 import {BaseInput, BaseTextArea} from 'src/components/assets/inputs';
 import {PrimaryButton} from 'src/components/assets/buttons';
 import {VerticalSpacer} from 'src/components/backstage/playbook_runs/shared';
 import {DollarSign, PoundSign} from 'src/components/backstage/playbook_edit/styles';
+import {stringToTarget, targetToString} from 'src/components/backstage/playbook_edit/metrics/shared';
 
 interface Props {
     metric: Metric;
@@ -146,43 +146,6 @@ const MetricEdit = ({metric, otherTitles, onAdd, saveToggle, saveFailed}: Props)
             </Container>
         </>
     );
-};
-
-const targetToString = (target: number, type: MetricType) => {
-    if (!target) {
-        if (type === MetricType.Integer || type === MetricType.Currency) {
-            return '0';
-        }
-        return '00:00:00';
-    }
-
-    if (type === MetricType.Integer || type === MetricType.Currency) {
-        return target.toString();
-    }
-
-    const dur = Duration.fromMillis(target).shiftTo('days', 'hours', 'minutes');
-    const dd = dur.days.toString().padStart(2, '0');
-    const hh = dur.hours.toString().padStart(2, '0');
-    const mm = dur.minutes.toString().padStart(2, '0');
-    return `${dd}:${hh}:${mm}`;
-};
-
-const stringToTarget = (target: string, type: MetricType) => {
-    if (target === '') {
-        return 0;
-    }
-
-    if (type === MetricType.Integer || type === MetricType.Currency) {
-        return parseInt(target, 10);
-    }
-
-    // assuming we've verified this is a duration in the format dd:mm:ss
-    const ddmmss = target.split(':').map((c) => parseInt(c, 10));
-    return Duration.fromObject({
-        days: ddmmss[0],
-        hours: ddmmss[1],
-        minutes: ddmmss[2],
-    }).as('milliseconds');
 };
 
 const Header = styled.div`
