+++ webapp/src/components/backstage/playbook_edit/metrics/metrics.tsx
@@ -8,8 +8,8 @@ import {useIntl} from 'react-intl';
 import {TertiaryButton} from 'src/components/assets/buttons';
 import DotMenu, {DropdownMenuItem} from 'src/components/dot_menu';
 import {DraftPlaybookWithChecklist, Metric, MetricType, newMetric, PlaybookWithChecklist} from 'src/types/playbook';
-import MetricEdit from 'src/components/backstage/playbook_edit/metric_edit';
-import MetricView from 'src/components/backstage/playbook_edit/metric_view';
+import MetricEdit from 'src/components/backstage/playbook_edit/metrics/metric_edit';
+import MetricView from 'src/components/backstage/playbook_edit/metrics/metric_view';
 import {DollarSign, PoundSign} from 'src/components/backstage/playbook_edit/styles';
 
 enum TaskType {
